content/base/test/test_bug804395.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=804395
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug 804395</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=804395">Mozilla Bug 804395</a>
<p id="display"></p>
<div id="content" style="display: none">

</div>
<pre id="test">
<script type="application/javascript">

function test200() {
  var xhr = new XMLHttpRequest({mozAnon: false, mozSystem: true});
  xhr.open('GET', 'jar:http://example.org/tests/content/base/test/file_bug804395.jar!/foo.bar', true);
  xhr.onreadystatechange = function() {
    if (xhr.readyState == 4) {
      ok(xhr.status == 200, "Existing file must have Status 200!");
      runTests();
    }
  }
  xhr.send(null);
}

function test404() {
  var xhr = new XMLHttpRequest({mozAnon: false, mozSystem: true});
  xhr.open('GET', 'jar:http://example.org/tests/content/base/test/file_bug804395.jar!/foo.do_not_exist', true);
  xhr.onreadystatechange = function() {
    if (xhr.readyState == 4) {
      ok(xhr.status == 404, "Non existing file must have Status 404!");
      runTests();
    }
  }
  xhr.send(null);
}

function test0() {
  var xhr = new XMLHttpRequest({mozAnon: false, mozSystem: true});
  xhr.open('GET', 'jar:http://example.org/tests/content/base/test/file_bug804395.jar!/foo.bar', true);
  ok(xhr.status == 0, "Not Sent request must have status 0");
  runTests();
}

var tests = [ test200, test404, test0 ];
function runTests() {
  if (!tests.length) {
    SimpleTest.finish();
    return;
  }

  var test = tests.shift();
  test();
}

/** Test for Bug 804395 **/
SimpleTest.waitForExplicitFinish();

addLoadEvent(function() {
   SpecialPowers.pushPermissions([{'type': 'systemXHR', 'allow': true, 'context': document}], runTests);
});
</script>
</pre>
</body>
</html>
